refactor(domain): Lotto 클래스 수정
* Lotto.java, WinningLotto.java: 문제 조건에 따라, Lotto 클래스를 사용하기 위해, 추상클래스였던, BasicLotto.java 를 삭제하고 메서드 구현
