+++ docs/수정_및_고민_사항.md
@@ -389,3 +389,10 @@ public class WinningLotto extends BasicLotto {
 * ``System.lineSeparator()`` 사용하여 해결
 
 ----
+## 수정사항
+
+**문제점**
+* Basic 로또를 추상클래스로 구현하여 Lotto와 WinningLotto를 구현하였지만, 문제 조건에 따라, 주어진 Lotto 클래스를 사용하기 위해 다시 재수정 할 필요 있음.
+
+**해결**
+* ``System.lineSeparator()`` 사용하여 해결
